+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.1.3] - 2020-07-30
+
+## Added
+
+- spec test for Debian install
+
+### Fixed
+
+- Typos in service.pp affecting Debian systems
+- Legacy facts replaced with structured facts
+
 ## [v2.1.2] - 2020-07-29
 
 ### Fixed